+++ vite.config.ts
@@ -1,0 +1,14 @@
+import { crx } from '@crxjs/vite-plugin';
+import { defineConfig } from 'vite';
+
+import manifest from './manifest.config';
+
+export default defineConfig({
+  plugins: [crx({ manifest })],
+  server: {
+    host: true,
+    port: 5173,
+    strictPort: true,
+    cors: true
+  }
+});